moved context tests into their own file
